requirements.txt
openai-whisper
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
pydub
static-ffmpeg
